+++ src/core/maps/js/mapMarker.js
@@ -83,13 +83,15 @@ function createPolygonMarker (styleId = "defaultMapMarkerPolygon") {
  * @returns {void}
  */
 function addFeatureToMapMarkerLayer (layerId, feature) {
-    const markerLayer = getMapmarkerLayerById(layerId),
-        styleId = markerLayer.get("styleId"),
-        styleObject = styleList.returnStyleObject(styleId),
-        featureStyle = createStyle.createStyle(styleObject, feature, false, Config.wfsImgPath);
+    if (getMapmarkerLayerById(layerId) !== undefined) {
+        const markerLayer = getMapmarkerLayerById(layerId),
+            styleId = markerLayer.get("styleId"),
+            styleObject = styleList.returnStyleObject(styleId),
+            featureStyle = createStyle.createStyle(styleObject, feature, false, Config.wfsImgPath);
 
-    feature.setStyle(featureStyle);
-    markerLayer.getSource().addFeature(feature);
+        feature.setStyle(featureStyle);
+        markerLayer.getSource().addFeature(feature);
+    }
 }
 
 /**