+++ src/modules/login/components/LoginComponent.vue
@@ -1,5 +1,5 @@
 <script>
-import {mapMutations, mapGetters, mapActions} from "vuex";
+import {mapGetters, mapActions} from "vuex";
 import {translateKeyWithPlausibilityCheck} from "../../../shared/js/utils/translateKeyWithPlausibilityCheck.js";
 
 export default {
@@ -10,14 +10,12 @@ export default {
         ...mapGetters(["isMobile"]),
         ...mapGetters("Modules/Login", ["loggedIn", "screenName", "email", "iconLogin", "iconLogged"])
     },
-    mounted () {
-        if (!this.isLoggedIn()) {
+    async mounted () {
+        if (!await this.isLoggedIn()) {
             this.openLoginWindow();
         }
     },
     methods: {
-        ...mapMutations("Menu", ["setCurrentComponentPropsName", "setCurrentComponentPropsDescription"]),
-        ...mapMutations("Modules/Login", ["setActive", "setIcon"]),
         ...mapActions("Modules/Login", [
             "initialize",
             "logout",
@@ -47,9 +45,8 @@ export default {
          * Returns true if user is logged in, else false
          * @return {Boolean} logged in
          */
-        isLoggedIn () {
-            this.checkLoggedIn();
-            this.setLoginProps();
+        async isLoggedIn () {
+            await this.checkLoggedIn();
             return this.loggedIn;
         },
 
@@ -97,26 +94,9 @@ export default {
         logoutButton (reload = false) {
             this.logout();
 
-            this.setLoginProps();
-
             if (reload) {
                 this.reloadWindow();
             }
-        },
-
-
-        /**
-         * Sets the icon, name and description according to the login status.
-         * @return {void}
-         */
-        setLoginProps () {
-            const iconType = this.loggedIn ? this.iconLogged : this.iconLogin,
-                componentName = this.loggedIn ? "common:modules.login.logout" : "common:modules.login.login",
-                componentDescription = this.loggedIn ? "common:modules.login.descriptionLoggedIn" : "common:modules.login.description";
-
-            this.setIcon(iconType);
-            this.setCurrentComponentPropsName({side: "secondaryMenu", name: componentName});
-            this.setCurrentComponentPropsDescription({side: "secondaryMenu", description: componentDescription});
         }
     }
 };