+++ src/modules/login/store/actionsLogin.js
@@ -67,33 +67,30 @@ export default {
      * @param {Object} context the context Vue instance
      * @return {Boolean} logged in
      */
-    checkLoggedIn ({commit, dispatch}) {
-
+    async checkLoggedIn ({commit, dispatch}) {
         const config = Config.login,
             token = Cookie.get("token"),
-            refreshToken = Cookie.get("refresh_token");
+            refreshToken = Cookie.get("refresh_token"),
+            loggedIn = Boolean(token);
 
-        let loggedIn = false;
-
+        // Set login props immediately based on token presence
+        commit("setIcon", loggedIn ? "bi-door-closed" : "bi-door-open");
+        commit("setName", loggedIn ? "common:modules.login.logout" : "common:modules.login.login");
 
         commit("setAccessToken", token);
         commit("setRefreshToken", refreshToken);
 
         if (OIDC.getTokenExpiry(token) < 1) {
-            dispatch("logout");
+            await dispatch("logout");
             return false;
         }
 
-        OIDC.renewTokenIfNecessary(token, refreshToken, config);
-
-        loggedIn = Boolean(token);
+        await OIDC.renewTokenIfNecessary(token, refreshToken, config);
 
         commit("setLoggedIn", loggedIn);
-
         commit("setScreenName", Cookie.get("name"));
         commit("setUsername", Cookie.get("username"));
         commit("setEmail", Cookie.get("email"));
-
         return loggedIn;
     }
 };